3932 update data loader (#974)
* data-reset-tool works again


* new pm collection


* updated posts for 2 bcomps

+++ data-reset-tool/src/data_reset_tool/blueprints/fixture.py
@@ -1,4 +1,5 @@
 """Endpoints for importing, exporting, and clearing business data."""
+import copy
 import io
 from http import HTTPStatus
 from zipfile import ZipFile
@@ -10,13 +11,59 @@ from flask import Blueprint, current_app, jsonify, request, send_file
 
 FIXTURE_BLUEPRINT = Blueprint('fixture', __name__)
 
+ALL_BCOMP_TABLES = \
+    [
+        'aliases_version',
+        'aliases',
+        'resolutions_version',
+        'resolutions',
+        'share_series_version',
+        'share_series',
+        'share_classes_version',
+        'share_classes',
+        'party_roles_version',
+        'party_roles',
+        'parties_version',
+        'parties',
+        'parties_version-addresses_version',
+        'parties_version-addresses',
+        'parties-addresses_version',
+        'parties-addresses',
+        'addresses_version',
+        'addresses',
+        'offices_version',
+        'offices',
+        'filings',
+        'businesses_version',
+        'businesses',
+        'transaction'
+    ]
 
-@FIXTURE_BLUEPRINT.route('/api/fixture/import', methods=['POST'], strict_slashes=False)
-@FIXTURE_BLUEPRINT.route('/api/fixture/import/<table>', methods=['POST'], strict_slashes=False)
-def post(table=None):
+ALL_COOP_TABLES = \
+    [
+        'party_roles_version',
+        'party_roles',
+        'parties-addresses_version',
+        'parties-addresses',
+        'parties_version',
+        'parties',
+        'addresses_version',
+        'addresses',
+        'offices_version',
+        'offices',
+        'filings',
+        'businesses_version',
+        'businesses',
+        'transaction'
+    ]
+
+
+@FIXTURE_BLUEPRINT.route('/api/fixture/import/<legal_type>', methods=['POST'], strict_slashes=False)
+@FIXTURE_BLUEPRINT.route('/api/fixture/import/<legal_type>/<table>', methods=['POST'], strict_slashes=False)
+def post(legal_type, table=None):
     """Import csv data into given table."""
     if not request.files.keys():
-        return jsonify({'message': f'No csv files given for import.'}), HTTPStatus.BAD_REQUEST
+        return jsonify({'message': 'No csv files given for import.'}), HTTPStatus.BAD_REQUEST
     # data to import
     try:
         # connection to db
@@ -28,31 +75,43 @@ def post(table=None):
             ), HTTPStatus.INTERNAL_SERVER_ERROR
         cur = con.cursor()
 
-        csv_list = [f'{table}'] if table \
-            else ['businesses', 'filings', 'offices', 'addresses', 'parties-addresses', 'parties', 'party_roles']
-
+        if table:
+            csv_list = [f'{table}']
+        elif legal_type == 'BC':
+            csv_list = copy.deepcopy(ALL_BCOMP_TABLES)
+        else:
+            csv_list = copy.deepcopy(ALL_COOP_TABLES)
+        csv_list.reverse()
         for filename in csv_list:
-            table = 'addresses' if filename == 'parties-addresses' else filename
-            input_file = request.files[f'{filename}']
-            try:
-                # delete existing entries
-                csv_data = pandas.read_csv(input_file)
-                ids = str(csv_data.id.to_list())
-                ids = ids.replace('[', '(').replace(']', ')')
-                if table == 'addresses':
-                    cur.execute(f'update parties set delivery_address_id=null where delivery_address_id in {ids}')
+            table = filename.replace('parties_version-', '').replace('parties-', '')
+            input_file = request.files.get(f'{filename}')
+            if input_file:
+                try:
+                    # delete existing entries
+                    csv_data = pandas.read_csv(input_file)
+                    ids = str(csv_data.id.to_list())
+                    ids = ids.replace('[', '(').replace(']', ')')
+                    if table in ['addresses', 'addresses_version']:
+                        cur.execute(
+                            f'update parties set delivery_address_id=null, mailing_address_id=null \
+                                where delivery_address_id in {ids} or mailing_address_id in {ids}'
+                        )
+                        cur.execute(
+                            f'update parties_version set delivery_address_id=null, mailing_address_id=null \
+                                where delivery_address_id in {ids} or mailing_address_id in {ids}'
+                        )
 
-                cur.execute(f'delete from {table} where id in {ids}')
-                if 'DELETE' not in cur.statusmessage:
-                    current_app.logger.error('Delete command did not run.')
-                    raise Exception
+                    cur.execute(f'delete from {table} where id in {ids}')
+                    if 'DELETE' not in cur.statusmessage:
+                        current_app.logger.error('Delete command did not run.')
+                        raise Exception
 
-            except Exception as err:
-                current_app.logger.error(f'Failed to delete existing entries: {err}')
+                except Exception as err:
+                    current_app.logger.error(f'Failed to delete existing entries: {err}')
 
-            # set reader back to beginning of file and import csv into table
-            input_file.seek(0)
-            cur.copy_expert(f"COPY {table} from stdin delimiter ',' csv header", input_file)
+                # set reader back to beginning of file and import csv into table
+                input_file.seek(0)
+                cur.copy_expert(f"COPY {table} from stdin delimiter ',' csv header", input_file)
 
         con.commit()
         return jsonify({'message': 'Success!'}), HTTPStatus.CREATED
@@ -87,8 +146,7 @@ def get(business_identifier, table=None):
             export_name = f'{business_identifier}-{table}'
             csv_files = _copy_from_table(cur=cur, table=table, business_id=business_id)
         else:
-            all_tables = ['businesses', 'filings', 'offices', 'party_roles', 'parties', 'addresses']
-            for item in all_tables:
+            for item in ALL_BCOMP_TABLES:
                 csv_files += _copy_from_table(cur=cur, table=item, business_id=business_id)
         if not csv_files:
             return jsonify(
@@ -129,21 +187,79 @@ def delete(business_identifier):
             current_app.logger.error(f'{business_identifier} not found.')
             return jsonify({'message': f'Could not find {business_identifier}.'}), HTTPStatus.NOT_FOUND
 
+        share_class_ids = _get_id_list(
+            cur=cur, column='id', table='share_classes', table_val='business_id', val=business_id)
+        share_class_version_ids = _get_id_list(
+            cur=cur, column='id', table='share_classes_version', table_val='business_id', val=business_id)
         party_ids = _get_id_list(
-            cur=cur, column='party_id', table='party_roles', table_val='business_id', val=business_id
-        )
+            cur=cur, column='party_id', table='party_roles', table_val='business_id', val=business_id)
         office_ids = _get_id_list(cur=cur, column='id', table='offices', table_val='business_id', val=business_id)
         filing_ids = _get_id_list(cur=cur, column='id', table='filings', table_val='business_id', val=business_id)
-
+        transaction_ids = _get_id_list(
+            cur=cur, column='transaction_id', table='filings', table_val='business_id', val=business_id)
         # delete all data for given business
         cur.execute(
             f"""
+            delete from party_roles_version where business_id={business_id};
             delete from party_roles where business_id={business_id};
-            delete from parties where id in {party_ids};
-            delete from addresses where office_id in {office_ids};
+            """
+        )
+        if _share_series_exists(cur, 'share_series_version', share_class_ids):
+            cur.execute(
+                f"""
+                delete from share_series_version where share_class_id in {share_class_ids};
+                """
+            )
+        if _share_series_exists(cur, 'share_series_version', share_class_version_ids):
+            cur.execute(
+                f"""
+                delete from share_series_version where share_class_id in {share_class_ids};
+                """
+            )
+        if _share_series_exists(cur, 'share_series', share_class_ids):
+            cur.execute(
+                f"""
+                delete from share_series where id in {share_class_ids};
+                """
+            )
+        if party_ids:
+            cur.execute(
+                f"""
+                delete from parties_version where id in {party_ids};
+                delete from parties where id in {party_ids};
+                """
+            )
+        if office_ids:
+            cur.execute(
+                f"""
+                delete from addresses_version where office_id in {office_ids};
+                delete from addresses where office_id in {office_ids};
+                """
+            )
+        if filing_ids:
+            cur.execute(
+                f"""
+                delete from colin_event_ids where filing_id in {filing_ids};
+                """
+            )
+        if transaction_ids:
+            cur.execute(
+                f"""
+                delete from filings where business_id={business_id};
+                delete from transaction where id in {transaction_ids};
+                """
+            )
+        cur.execute(
+            f"""
+            delete from aliases_version where business_id={business_id};
+            delete from aliases where business_id={business_id};
+            delete from resolutions_version where business_id={business_id};
+            delete from resolutions where business_id={business_id};
+            delete from share_classes_version where business_id={business_id};
+            delete from share_classes where business_id={business_id};
+            delete from offices_version where business_id={business_id};
             delete from offices where business_id={business_id};
-            delete from colin_event_ids where filing_id in {filing_ids};
-            delete from filings where id in {filing_ids};
+            delete from businesses_version where id={business_id};
             delete from businesses where id={business_id};
             """
         )
@@ -158,7 +274,7 @@ def delete(business_identifier):
         return jsonify({'message': f'Failed to delete {business_identifier}.'}), HTTPStatus.INTERNAL_SERVER_ERROR
 
 
-def _get_business_id(cur: psycopg2.extensions.cursor, business_identifier: str):
+def _get_business_id(cur: psycopg2.extensions.cursor, business_identifier: str) -> str:
     """Return the business id for the given identifier."""
     cur.execute(f"select id from businesses where identifier='{business_identifier}'")
     business_id = cur.fetchone()
@@ -167,59 +283,133 @@ def _get_business_id(cur: psycopg2.extensions.cursor, business_identifier: str):
     return str(business_id[0])
 
 
-def _get_id_list(cur: psycopg2.extensions.cursor, column: str, table: str, table_val: str, val: str):
+def _get_id_list(cur: psycopg2.extensions.cursor, column: str, table: str, table_val: str, val: str) -> str:
     """Return a stringified list of ids for the given table linked to the business_id."""
     val = val.replace('(', '').replace(')', '')
+    if not val:
+        return ''
     cur.execute(f'select {column} from {table} where {table_val} in ({val})')
     id_list = []
     for _id in cur.fetchall():
-        id_list.append(_id[0])
+        if _id[0]:
+            id_list.append(_id[0])
+    if not id_list:
+        return ''
     return str(id_list).replace('[', '(').replace(']', ')')
 
 
-def _create_csv(cur: psycopg2.extensions.cursor, filename: str, select_stmnt: str):
+def _create_csv(cur: psycopg2.extensions.cursor, filename: str, select_stmnt: str) -> str:
     """Create or update given csv file with db data for given select statement."""
     with open(f'exports/{filename}.csv', 'wb') as csvfile:
         cur.copy_expert(f'COPY ({select_stmnt}) to stdout with csv header', csvfile)
     return f'{filename}.csv'
 
 
-def _copy_from_table(cur: psycopg2.extensions.cursor, table: str, business_id: str):
+def _copy_from_table(cur: psycopg2.extensions.cursor, table: str, business_id: str) -> list:
     """Copy db data into csv files for given table."""
     files = []
     select_stmnt = ''
-    if table in ['filings', 'offices', 'party_roles']:
+
+    if table in [
+        'aliases_version',
+        'aliases',
+        'resolutions_version',
+        'resolutions',
+        'share_classes_version',
+        'share_classes',
+        'party_roles_version',
+        'party_roles',
+        'offices_version',
+        'offices',
+        'filings'
+    ]:
         # copy from table based on business_id
         id_list = _get_id_list(cur=cur, column='id', table=table, table_val='business_id', val=business_id)
-        select_stmnt = f'select * from {table} where id in {id_list}'
+        if id_list:
+            select_stmnt = f'select * from {table} where id in {id_list}'
 
-    elif table == 'parties':
+    elif table in ['parties', 'parties_version']:
         # copy parties and their addresses based on party_id and delivery_address_id
         # get party ids from party roles table
         id_list = _get_id_list(
             cur=cur, column='party_id', table='party_roles', table_val='business_id', val=business_id
         )
-        # get address ids from parties table
-        addr_id_list = _get_id_list(
-            cur=cur, column='delivery_address_id', table='parties', table_val='id', val=id_list
-        )
-        # create address csv for party addresses
-        addresses_stmnt = f'select * from addresses where id in {addr_id_list}'
-        files.append(_create_csv(cur=cur, filename=f'{table}-addresses', select_stmnt=addresses_stmnt))
+        if id_list:
+            # get address ids from parties table
+            deliv_addr_id_list = _get_id_list(
+                cur=cur, column='delivery_address_id', table='parties', table_val='id', val=id_list)
+            # create address csv for party delivery addresses
+            if deliv_addr_id_list:
+                addresses_stmnt = f'select * from addresses where id in {deliv_addr_id_list}'
+                addresses_version_stmnt = f'select * from addresses_version where id in {deliv_addr_id_list}'
+                files.append(_create_csv(cur=cur, filename=f'{table}-addresses', select_stmnt=addresses_stmnt))
+                files.append(
+                    _create_csv(cur=cur, filename=f'{table}-addresses_version', select_stmnt=addresses_version_stmnt))
+            mail_addr_id_list = _get_id_list(
+                cur=cur, column='mailing_address_id', table='parties', table_val='id', val=id_list)
+            # create address csv for party mailing addresses
+            if mail_addr_id_list:
+                addresses_stmnt = f'select * from addresses where id in {mail_addr_id_list}'
+                addresses_version_stmnt = f'select * from addresses_version where id in {mail_addr_id_list}'
+                files.append(_create_csv(cur=cur, filename=f'{table}-addresses', select_stmnt=addresses_stmnt))
+                files.append(
+                    _create_csv(cur=cur, filename=f'{table}-addresses_version', select_stmnt=addresses_version_stmnt))
+            select_stmnt = f'select * from {table} where id in {id_list}'
 
-        select_stmnt = f'select * from {table} where id in {id_list}'
-
-    elif table == 'addresses':
+    elif table in ['addresses', 'addresses_version']:
         # copy from addresses based on offices ids
         id_list = _get_id_list(cur=cur, column='id', table='offices', table_val='business_id', val=business_id)
-        select_stmnt = f'select * from {table} where office_id in {id_list}'
+        if id_list:
+            select_stmnt = f'select * from {table} where office_id in {id_list}'
 
-    elif table == 'businesses':
+    elif table in ['share_series', 'share_series_version']:
+        # copy from share series based on share classes ids
+        share_class_version_ids = _get_id_list(
+            cur=cur, column='id', table='share_classes_version', table_val='business_id', val=business_id)
+        share_class_ids = _get_id_list(
+            cur=cur, column='id', table='share_classes', table_val='business_id', val=business_id)
+        # only create this select statement if there is a valid share series
+        if table == 'share_series_version':
+            if _share_series_exists(cur, table, share_class_ids) or _share_series_exists(
+                cur, table, share_class_version_ids
+            ):
+                select_stmnt = f'select * from {table} where share_class_id in {share_class_ids}'
+        elif _share_series_exists(cur, table, share_class_ids):
+            select_stmnt = f'select * from {table} where share_class_id in {share_class_ids}'
+
+    elif table in ['transaction']:
+        # copy from transactions based on filing transaction ids
+        id_list = _get_id_list(
+            cur=cur, column='transaction_id', table='filings', table_val='business_id', val=business_id)
+        if id_list:
+            select_stmnt = f'select * from {table} where id in {id_list}'
+
+    elif table in ['businesses', 'businesses_version']:
         # copy from businesses table
         select_stmnt = f'select * from {table} where id={business_id}'
-    else:
-        current_app.logger.error(f'No export built for {table}.')
-        return []
 
-    files.append(_create_csv(cur=cur, filename=table, select_stmnt=select_stmnt))
+    else:
+        # otherwise handled elsewhere
+        if table not in [
+            'parties-addresses_version',
+            'parties-addresses',
+            'parties_version-addresses_version',
+            'parties_version-addresses'
+        ]:
+            current_app.logger.error(f'No export built for {table}.')
+
+    if select_stmnt:
+        files.append(_create_csv(cur=cur, filename=table, select_stmnt=select_stmnt))
+
     return files
+
+
+def _share_series_exists(cur: psycopg2.extensions.cursor, table: str, share_class_ids: list) -> bool:
+    share_series = None
+    for share_class_id in share_class_ids:
+        share_series = _get_id_list(
+            cur=cur, column='id', table=f'{table}', table_val='share_class_id', val=share_class_id)
+        if share_series:
+            break
+
+    return share_series is not None
